Enable "Immersive-AR" session for OpenXR runtime
The OpenXR runtime on Windows supports both "opaque" (WMR VR devices) and "additive" (Hololens). This change simply turns on support for Immersive-AR session according to the WebXR AR-module spec. The majority of the change is to let the runtime to decide about the blend_mode and interaction_mode instead of the current guessing logic at at the renderer layer.

Bug: 1069350, 944936
Change-Id: Ie5f1b56857f9ca5357dbd1dde6585687773d0b56